Rename models to entities
This more accurately reflects the meaning.
